src/requirements.txt
python-telegram-bot
pymongo
bson
configparser
requests
